Started work on decompositions.
