+++ src/Simulation/Simulators/OpenSystemsSimulator/OpenSystemsSimulator.cs
@@ -41,56 +41,9 @@ namespace Microsoft.Quantum.Experimental
 
         public State CurrentState => NativeInterface.GetCurrentState(this.Id);
 
-        public class OpenSystemsQubitManager : QubitManager
-        {
-            private readonly OpenSystemsSimulator Parent;
-            public OpenSystemsQubitManager(OpenSystemsSimulator parent, uint capacity)
-                : base(capacity)
-            {
-                this.Parent = parent;
-            }
-
-            protected override void Release(Qubit qubit, bool wasUsedOnlyForBorrowing)
-            {
-                if (qubit != null && qubit.IsMeasured)
-                {
-                    // Try to reset measured qubits.
-                    // TODO: There are better ways to do this; increment on the
-                    //       design and make it customizable.
-                    // FIXME: In particular, this implementation uses a lot of
-                    //        extraneous measurements.
-                    if ((this.Parent as IIntrinsicMeasure).Body(new QArray<Pauli>(Pauli.PauliZ), new QArray<Qubit>(qubit)) == Result.One)
-                    {
-                        (this.Parent as IIntrinsicX).Body(qubit);
-                    }
-                }
-                base.Release(qubit, wasUsedOnlyForBorrowing);
-            }
-        }
-
         public OpenSystemsSimulator(uint capacity = 3, string representation = "mixed") : base(new QubitManager((long)capacity))
         {
             this.Id = NativeInterface.Init(capacity, representation);
-        }
-
-        void IIntrinsicExp.Body(IQArray<Pauli> paulis, double angle, IQArray<Qubit> targets)
-        {
-            throw new NotImplementedException();
-        }
-
-        void IIntrinsicExp.AdjointBody(IQArray<Pauli> paulis, double angle, IQArray<Qubit> targets)
-        {
-            throw new NotImplementedException();
-        }
-
-        void IIntrinsicExp.ControlledBody(IQArray<Qubit> controls, IQArray<Pauli> paulis, double angle, IQArray<Qubit> targets)
-        {
-            throw new NotImplementedException();
-        }
-
-        void IIntrinsicExp.ControlledAdjointBody(IQArray<Qubit> controls, IQArray<Pauli> paulis, double angle, IQArray<Qubit> targets)
-        {
-            throw new NotImplementedException();
         }
 
         void IIntrinsicH.Body(Qubit target)
@@ -100,7 +53,14 @@ namespace Microsoft.Quantum.Experimental
 
         void IIntrinsicH.ControlledBody(IQArray<Qubit> controls, Qubit target)
         {
-            throw new NotImplementedException();
+            if (controls is { Count: 0 })
+            {
+                (this as IIntrinsicH).Body(target);
+            }
+            else
+            {
+                Get<Decompositions.ApplyControlledH, Decompositions.ApplyControlledH>().__Body__((controls, target));
+            }
         }
 
         Result IIntrinsicMeasure.Body(IQArray<Pauli> paulis, IQArray<Qubit> targets)
@@ -111,41 +71,8 @@ namespace Microsoft.Quantum.Experimental
             }
             else
             {
-                // FIXME: Pass multiqubit and non-Z cases to decompositions.
-                throw new NotImplementedException();
+                Get<Decompositions.Measure, Decompositions.Measure>().__Body__((paulis, targets));
             }
-        }
-
-        void IIntrinsicR.Body(Pauli pauli, double angle, Qubit target)
-        {
-            if (pauli == Pauli.PauliI)
-            {
-                // Don't apply global phases on uncontrolled operations.
-                return;
-            }
-            throw new NotImplementedException("Arbitrary rotation with noise is not yet supported.");
-        }
-
-        void IIntrinsicR.AdjointBody(Pauli pauli, double angle, Qubit target)
-        {
-            (this as IIntrinsicR).Body(pauli, -angle, target);
-        }
-
-        void IIntrinsicR.ControlledBody(IQArray<Qubit> controls, Pauli pauli, double angle, Qubit target)
-        {
-            if (controls is { Count: 0 })
-            {
-                (this as IIntrinsicR).Body(pauli, angle, target);
-            }
-            else
-            {
-                throw new NotImplementedException("Arbitrary controlled rotation with noise is not yet supported.");
-            }
-        }
-
-        void IIntrinsicR.ControlledAdjointBody(IQArray<Qubit> controls, Pauli pauli, double angle, Qubit target)
-        {
-            (this as IIntrinsicR).ControlledBody(controls, pauli, -angle, target);
         }
 
         void IIntrinsicS.Body(Qubit target)
@@ -234,7 +161,7 @@ namespace Microsoft.Quantum.Experimental
             }
             else if (controls is { Count: 1 })
             {
-                Get<ApplyCZUsingCNOT, ApplyCZUsingCNOT>().__Body__((controls[0], target));
+                Get<Decompositions.ApplyCZUsingCNOT, Decompositions.ApplyCZUsingCNOT>().__Body__((controls[0], target));
             }
             else
             {
@@ -247,43 +174,5 @@ namespace Microsoft.Quantum.Experimental
             NativeInterface.Destroy(this.Id);
         }
 
-        public class OpenSystemsDumpMachine<T> : Quantum.Diagnostics.DumpMachine<T>
-        {
-            private OpenSystemsSimulator Simulator { get; }
-
-            public OpenSystemsDumpMachine(OpenSystemsSimulator m) : base(m)
-            {
-                this.Simulator = m;
-            }
-
-            public override Func<T, QVoid> __Body__ => (location) =>
-            {
-                if (location == null) { throw new ArgumentNullException(nameof(location)); }
-                Simulator.MaybeDisplayDiagnostic(Simulator.CurrentState);
-                return QVoid.Instance;
-            };
-        }
-
-        // TODO: implement this by adding a new PartialTrace trait to the
-        //       Rust side, and then exposing it through the C API.
-        //       Until we have that, there's not a sensible way to interpret
-        //       states on subregisters in general.
-        public class OpenSystemsDumpRegister<T> : Quantum.Diagnostics.DumpRegister<T>
-        {
-            private OpenSystemsSimulator Simulator { get; }
-
-            public OpenSystemsDumpRegister(OpenSystemsSimulator m) : base(m)
-            {
-                this.Simulator = m;
-            }
-
-            public override Func<(T, IQArray<Qubit>), QVoid> __Body__ => (args) =>
-            {
-                var (location, register) = args;
-                if (location == null) { throw new ArgumentNullException(nameof(location)); }
-                this.Simulator.Get<Message, Message>().__Body__?.Invoke("DumpRegister not yet supported on OpenSystemsSimulator; skipping.");
-                return QVoid.Instance;
-            };
-        }
     }
 }
